+++ src/app/components/general/gen-arrows/gen-arrows.component.scss
@@ -1,4 +1,4 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
 
 /* Only NOT Mobile */
 @media only screen and (hover: hover) and (pointer: fine) {
