+++ src/app/components/general/gen-info/gen-info.component.scss
@@ -1,4 +1,4 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
 
 .info-back {
   position: fixed;
