+++ src/app/components/navigators/nav-footer/nav-footer.component.scss
@@ -1,4 +1,4 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
 
 .footer {
   border-color: $color-purple !important;
