+++ src/app/components/navigators/nav-header/nav-header.component.scss
@@ -1,4 +1,4 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
 
 .navbar {
   padding: 0px;
