+++ src/app/components/screens/scr-about/scr-about.component.scss
@@ -1,4 +1,4 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
 
 .content-block {
   height: auto;
