+++ src/app/components/screens/scr-curriculum/scr-curriculum.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
-import { commonModuleUtils } from 'src/app/utils/RoutingUtils';
-import { ScreenSupport, Screens } from 'src/app/utils/ScreenUtils';
+import { commonModuleUtils } from 'src/app/utils/routing.utils';
+import { ScreenSupport, Screens } from 'src/app/utils/screen.utils';
 
 import { GenArrowsComponent } from '../../general/gen-arrows/gen-arrows.component';
 import { GenInfoComponent } from '../../general/gen-info/gen-info.component';
