+++ src/app/components/screens/scr-info/scr-info.component.scss
@@ -1,4 +1,4 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
 
 b {
   color: #fff;
