+++ src/app/components/screens/scr-info/scr-info.component.ts
@@ -1,8 +1,6 @@
 import { Component } from '@angular/core';
 
-import { commonModuleUtils } from 'src/app/utils/RoutingUtils';
-import { ScreenSupport } from 'src/app/utils/ScreenUtils';
-import { Screens } from 'src/app/utils/ScreenUtils';
+import { commonModuleUtils } from 'src/app/utils/routing.utils';
 
 import { GenArrowsComponent } from '../../general/gen-arrows/gen-arrows.component';
 import { GenInfoComponent } from '../../general/gen-info/gen-info.component';
