+++ src/app/components/screens/scr-knowledge/scr-knowledge.component.scss
@@ -1,4 +1,4 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
 
 h6 {
   color: $color-purple;
