+++ src/app/components/screens/scr-main/scr-main.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
-import { commonModuleUtils } from 'src/app/utils/RoutingUtils';
-import { ScreenSupport, Screens } from 'src/app/utils/ScreenUtils';
+import { commonModuleUtils } from 'src/app/utils/routing.utils';
+import { ScreenSupport, Screens } from 'src/app/utils/screen.utils';
 
 import { GenArrowsComponent } from '../../general/gen-arrows/gen-arrows.component';
 import { GenDeviconsComponent } from '../../general/gen-devicons/gen-devicons.component';
