+++ src/app/components/screens/scr-skills/scr-skills.component.scss
@@ -1,1 +1,1 @@
-@import '../../../utils/StyleUtils.scss';
+@import '../../../utils/style.utils.scss';
