+++ src/styles.scss
@@ -1,4 +1,4 @@
-@import '../src/app/utils/StyleUtils.scss';
+@import '../src/app/utils/style.utils.scss';
 
 // Default Elements
